Move replication.tcp.streams into a package
